[Yapsi] expose the runtime state a bit more
This commit is explicitly tailored to expose the parts that Tardis will need
in order to register the state at each tick.

The four calls to self.*tick (beginning of program, after assignment, after
binding, after call to &say) are in a way arbitrary, but it's the best
guess we have right now as to what a debugger needs. The reason we use a
* rather than a ? in the call is so that many deriving classes will be able
to have their .tick method invoked. Will be interesting to see if that
distinction will actually be important to someone.