re-order things in expression parsing
I don't see how the present order could ever have worked. If we parse a
variable before we parse an invocation (which is defined as variable-or-block
and then '()'), then the <variable> rule will always succeed where the
<invocation> rule would have succeeded.

Of course, all this is screaming for a proper expression parser rather than
the temporary/fragile structure that's there now.